Added evaluator, implemented dealer chip division algorithm

+++ src/main/java/game/evaluator/Evaluator.java
@@ -6,5 +6,5 @@ import game.player.Player;
 
 public interface Evaluator {
 
-	public Map<Player, Long> evaluatePlayers();
+	public Map<Player, Integer> evaluatePlayers();
 }
